add reliable-slc-configuration command
We use this command to reconfigure eMMC on beaglebones into their most
reliable configuration (SLC mode with reliable writes enabled).

Beware that executing this command will cause eMMC to

    PERMANENTLY LOSE 50% OF ITS STORAGE CAPACITY!

    THERE IS NO WAY TO UNDO THIS!

It also requires the beaglebone to be power-cycled afterwards, and
erases eMMC completely.  If eMMC is already in SLC mode with reliable
writes enabled, then the command simply does nothing.